+++ src/main/kotlin/com/walter/pos/repository/UserRepository.kt
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository
 import java.util.Optional
 
 @Repository
-interface UserRepository : JpaRepository<User, Int> {
+interface UserRepository : JpaRepository<User, Long> {
     fun findByUsername(username: String): Optional<User>
 }
