+++ src/main/kotlin/com/walter/pos/service/ProductUnitService.kt
@@ -14,7 +14,7 @@ class ProductUnitService(private val unitRepository: ProductUnitRepository) {
     fun getAllUnits(): List<ProductUnitResponse> =
         unitRepository.findAll().map { it.toResponse() }
 
-    fun getUnitById(id: Int): ProductUnitResponse =
+    fun getUnitById(id: Long): ProductUnitResponse =
         findUnitById(id).toResponse()
 
     fun createUnit(request: ProductUnitRequest): ProductUnitResponse {
@@ -22,14 +22,14 @@ class ProductUnitService(private val unitRepository: ProductUnitRepository) {
         return unitRepository.save(unit).toResponse()
     }
 
-    fun updateUnit(id: Int, request: ProductUnitRequest): ProductUnitResponse {
+    fun updateUnit(id: Long, request: ProductUnitRequest): ProductUnitResponse {
         val unit = findUnitById(id)
         unit.name = request.name
         unit.shortName = request.shortName
         return unitRepository.save(unit).toResponse()
     }
 
-    fun deleteUnit(id: Int) {
+    fun deleteUnit(id: Long) {
         if (!unitRepository.existsById(id)) {
             throw ResourceNotFoundException("Unit with ID $id not found.")
         }
@@ -38,7 +38,7 @@ class ProductUnitService(private val unitRepository: ProductUnitRepository) {
     fun searchUnits(query: String): List<ProductUnitResponse> =
         unitRepository.search(query).map { it.toResponse() }
 
-    internal fun findUnitById(id: Int): ProductUnit =
+    internal fun findUnitById(id: Long): ProductUnit =
         unitRepository.findById(id)
             .orElseThrow { ResourceNotFoundException("Unit with ID $id not found.") }
